+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Next release
 
+* `recompose` is no longer a dependency on any of the packages.
+* `calendar` package has been refactored and no longer depends on `recompose`.
 * Clickable forwards DOM attribute props, such as id and className.
 * Build process includes imports checker, that checks that imported modules are specified as dependency in package.json.
 * Add `useDomId` hook to `core`. Generates a unique id, that is guaranteed to persist through component lifecycle.