+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add first version of `@stenajs-webui/tooltip`.
 * Add `WithTooltip` component.
+* Add `MultiDateCalendar` component. User can select range as well as toggle individual dates.
 * Add standardized sizes to `Progress`, `small`, `normal`, `large`.
 * `Progress` `size` prop now sets height (previously only set width).
 * `recompose` is no longer a dependency on any of the packages.