+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 * Add `WithTooltip` component.
 * Add `MultiDateCalendar` component. User can select range as well as toggle individual dates.
 * Add standardized sizes to `Progress`, `small`, `normal`, `large`.
+* Add `invalid` prop to `StandardTextInput` with updated theme.
 * `Progress` `size` prop now sets height (previously only set width).
 * `recompose` is no longer a dependency on any of the packages.
 * `calendar` package has been refactored and no longer depends on `recompose`.
 * Clickable forwards DOM attribute props, such as id and className.
 * Build process includes imports checker, that checks that imported modules are specified as dependency in package.json.
 * Add `useDomId` hook to `core`. Generates a unique id, that is guaranteed to persist through component lifecycle.
+* Fix eslint in build process.
 
 #### 0.0.32